Factorization of Adresse related fields in a MappedSuperclass.
It has absolutely no side effects on DB schema.
