+++ src/mgate/PersonneBundle/Entity/Prospect.php
@@ -25,7 +25,7 @@ use Symfony\Component\Validator\Constraints as Assert;
  * @ORM\HasLifecycleCallbacks
  * @ORM\Entity(repositoryClass="mgate\PersonneBundle\Entity\ProspectRepository")
  */
-class Prospect
+class Prospect extends Adressable
 {
     /**
      * @var int
@@ -62,34 +62,6 @@ class Prospect
      * @Assert\Choice(callback = "getEntiteChoiceAssert")
      */
     private $entite;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="adresse", type="string", length=127, nullable=true)
-     */
-    private $adresse;
-
-    /**
-     * @var int(5)
-     *
-     * @ORM\Column(name="codepostal", type="integer", nullable=true)
-     */
-    private $codepostal;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="ville", type="string", length=63, nullable=true)
-     */
-    private $ville;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="pays", type="string", length=63, nullable=true)
-     */
-    private $pays;
 
     /**
      * Get id.
@@ -258,104 +230,4 @@ class Prospect
         return $tab[$this->entite];
     }
 
-    /**
-     * Set adresse.
-     *
-     * @param string $adresse
-     *
-     * @return Prospect
-     */
-    public function setAdresse($adresse)
-    {
-        $this->adresse = $adresse;
-
-        return $this;
-    }
-
-    /**
-     * Get adresse.
-     *
-     * @return string
-     */
-    public function getAdresse()
-    {
-        return $this->adresse;
-    }
-
-    /**
-     * Set codepostal.
-     *
-     * @param int $codepostal
-     *
-     * @return Prospect
-     */
-    public function setCodePostal($codepostal)
-    {
-        $this->codepostal = $codepostal;
-
-        return $this;
-    }
-
-    /**
-     * Get codePostal.
-     *
-     * @return int
-     */
-    public function getCodePostal()
-    {
-        return $this->codepostal;
-    }
-
-    /**
-     * Set ville.
-     *
-     * @param string $ville
-     *
-     * @return Prospect
-     */
-    public function setVille($ville)
-    {
-        $this->ville = $ville;
-
-        return $this;
-    }
-
-    /**
-     * Get ville.
-     *
-     * @return string
-     */
-    public function getVille()
-    {
-        return $this->ville;
-    }
-
-    /**
-     * Set codepostal.
-     *
-     * @param string $pays
-     *
-     * @return Prospect
-     */
-    public function setPays($pays)
-    {
-        $this->pays = $pays;
-
-        return $this;
-    }
-
-    /**
-     * Get pays.
-     *
-     * @return string
-     */
-    public function getPays()
-    {
-        return $this->pays;
-    }
-
-    public function __toString()
-    {
-        return $this->getNom();
-    }
 }
